+++ src/data/index.tsx
@@ -1,0 +1,60 @@
+// Dummy Data For Story Examples
+
+export const people = [
+  {
+    id: "0493240",
+    name: "Andrew Alfred",
+    role: "Technical advisor",
+    image:
+      "https://images.unsplash.com/photo-1501196354995-cbb51c65aaea?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "45945745",
+    name: "Debra Houston",
+    role: "Analyst",
+    image:
+      "https://images.unsplash.com/photo-1531123897727-8f129e1688ce?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "0689646",
+    name: "Jane White",
+    role: "Director, Marketing",
+    image:
+      "https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "86384385",
+    name: "Ray Flint",
+    role: "Technical Advisor",
+    image:
+      "https://images.unsplash.com/photo-1531427186611-ecfd6d936c79?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "86387385",
+    name: "Andrew Alfred",
+    role: "Technical advisor",
+    image:
+      "https://images.unsplash.com/photo-1501196354995-cbb51c65aaea?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "86364385",
+    name: "Debra Houston",
+    role: "Analyst",
+    image:
+      "https://images.unsplash.com/photo-1531123897727-8f129e1688ce?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "86364382",
+    name: "Jane White",
+    role: "Director, Marketing",
+    image:
+      "https://images.unsplash.com/photo-1517841905240-472988babdf9?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+  {
+    id: "86364395",
+    name: "Ray Flint",
+    role: "Technical Advisor",
+    image:
+      "https://images.unsplash.com/photo-1531427186611-ecfd6d936c79?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=facearea&facepad=4&w=256&h=256&q=80",
+  },
+];